generating bytecode for expressions

+++ include/noctis_script/ncsc.hpp
@@ -11,7 +11,7 @@ using usize_t   = std::uint64_t;
 using byte_t    = std::uint8_t;
 using word_t    = std::uint16_t;
 using dword_t   = std::uint32_t;
-using qword_t   = std::uint32_t;
+using qword_t   = std::uint64_t;
 
 using float32_t = float;
 using float64_t = double;
